+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to GOBL will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/) and this project adheres to [Semantic Versioning](http://semver.org/). See also the [GOBL versions](https://docs.gobl.org/overview/versions) documentation site for more details.
 
+## [vXXXX] - XXXX-XX-XX
+
+### Fixed
+
+- UUID `IsNotZero` will not raise error for empty UUIDs.
+
 ## [v0.74.1] - 2024-05-23
 
 UUID Unmarshal fix.